feat: Add storybook utils to labs common
