+++ modules/labs-react/common/lib/storybook-utils/propTypes.ts
@@ -1,0 +1,15 @@
+type PropDeclaration = {
+  value: any;
+  label: string;
+};
+
+type PropsDeclaration = {[key: string]: PropDeclaration[]};
+
+type Props = {
+  [key: string]: any;
+};
+
+type PropCombination = {
+  label: string;
+  props: Props;
+};
