+++ modules/labs-react/drawer/stories/stories_visualTesting.tsx
@@ -1,7 +1,9 @@
 /// <reference path="../../../../typings.d.ts" />
 import * as React from 'react';
 import {StaticStates} from '@workday/canvas-kit-react/common';
-import {ComponentStatesTable, withSnapshotsEnabled} from '../../../../utils/storybook';
+import {ComponentStatesTable} from '@workday/canvas-kit-labs-react/common';
+
+import {withSnapshotsEnabled} from '../../../../utils/storybook';
 import {space} from '@workday/canvas-kit-react/tokens';
 
 import {Drawer, DrawerHeader, DrawerDirection} from '../index';
