+++ modules/preview-react/breadcrumbs/stories/stories_visualTesting.tsx
@@ -3,7 +3,8 @@ import React from 'react';
 import withReadme from 'storybook-readme/with-readme';
 import {CanvasProvider, ContentDirection, StaticStates} from '@workday/canvas-kit-react/common';
 
-import {ComponentStatesTable, withSnapshotsEnabled} from '../../../../utils/storybook';
+import {ComponentStatesTable} from '@workday/canvas-kit-labs-react/common';
+import {withSnapshotsEnabled} from '../../../../utils/storybook';
 
 import {Breadcrumbs} from '../index';
 import README from '../README.md';
