+++ modules/react/badge/stories/stories_visualTesting.tsx
@@ -2,7 +2,8 @@
 import React from 'react';
 import withReadme from 'storybook-readme/with-readme';
 
-import {ComponentStatesTable, withSnapshotsEnabled} from '../../../../utils/storybook';
+import {ComponentStatesTable} from '@workday/canvas-kit-labs-react/common';
+import {withSnapshotsEnabled} from '../../../../utils/storybook';
 
 import {CountBadge} from '../index';
 import README from '../README.md';
