+++ modules/react/button/stories/visual-testing/stories_deprecatedButton.tsx
@@ -2,11 +2,8 @@
 /** @jsx jsx */
 import {jsx} from '@emotion/core';
 import {StaticStates} from '@workday/canvas-kit-react/common';
-import {
-  ComponentStatesTable,
-  permutateProps,
-  withSnapshotsEnabled,
-} from '../../../../../utils/storybook';
+import {ComponentStatesTable, permutateProps} from '@workday/canvas-kit-labs-react/common';
+import {withSnapshotsEnabled} from '../../../../../utils/storybook';
 import {deprecated_Button as DeprecatedButton} from '@workday/canvas-kit-react/button';
 import {Container, stateTableColumnProps} from './utils';
