+++ modules/react/form-field/stories/visual-testing/stories_Checkbox.tsx
@@ -1,12 +1,8 @@
 /// <reference path="../../../../../typings.d.ts" />
 import * as React from 'react';
 import {StaticStates} from '@workday/canvas-kit-react/common';
-import {
-  ComponentStatesTable,
-  permutateProps,
-  withSnapshotsEnabled,
-  customColorTheme,
-} from '../../../../../utils/storybook';
+import {ComponentStatesTable, permutateProps} from '@workday/canvas-kit-labs-react/common';
+import {withSnapshotsEnabled, customColorTheme} from '../../../../../utils/storybook';
 
 import {Checkbox} from '../../../checkbox';
