+++ modules/react/form-field/stories/visual-testing/stories_Switch.tsx
@@ -1,11 +1,7 @@
 /// <reference path="../../../../../typings.d.ts" />
 import * as React from 'react';
-import {
-  ComponentStatesTable,
-  permutateProps,
-  withSnapshotsEnabled,
-  customColorTheme,
-} from '../../../../../utils/storybook';
+import {ComponentStatesTable, permutateProps} from '@workday/canvas-kit-labs-react/common';
+import {withSnapshotsEnabled, customColorTheme} from '../../../../../utils/storybook';
 import {StaticStates} from '@workday/canvas-kit-react/common';
 import {Switch} from '../../../switch';
