+++ utils/create-component/templates/react/stories_VisualTesting.js
@@ -3,7 +3,8 @@
 module.exports = (modulePath, storyPath, pascalCaseName, rootPath) => `import React from 'react';
 
 import {StaticStates} from '@workday/canvas-kit-preview-react/tokens';
-import {ComponentStatesTable, withSnapshotsEnabled} from '../${rootPath}/utils/storybook';
+import {ComponentStatesTable} from '@workday/canvas-kit-labs-react/common';
+import {withSnapshotsEnabled} from '../${rootPath}/utils/storybook';
 
 import {${pascalCaseName}, use${pascalCaseName}Model} from '${modulePath}';
