+++ utils/storybook/index.ts
@@ -7,5 +7,4 @@ export {
 export {customColorTheme} from './customThemes';
 export {withSnapshotsEnabled} from './withSnapshotsEnabled';
 export {default as CanvasProviderDecorator} from './CanvasProviderDecorator';
-export {default as ComponentStatesTable, permutateProps} from './ComponentStatesTable';
 export * from './testingOnly';
